backend/requirements.txt
fastapi
uvicorn
torch
numpy
nltk
python-jose
passlib
python-multipart
